requirements.txt
beautifulsoup4==4.14.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
fonttools==4.60.1
frozendict==2.4.6
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.9
llvmlite==0.45.1
matplotlib==3.10.6
multitasking==0.0.12
mypy==1.18.2
mypy_extensions==1.1.0
numba==0.62.1
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
polars==1.34.0
polars-runtime-32==1.34.0
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
ruff==0.13.3
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.66
